Provide index file to export extracted modules
