+++ client/src/ClusterOverview.js
@@ -1,6 +1,6 @@
 import styles from './ClusterOverview.module.css';
 import { ClusterDescription, ClusterLogo } from './lib/Branding';
-import { useEnvironment } from './lib/BrandingContext';
+import { useEnvironment } from './lib';
 
 function ClusterName({ children }) {
   return <h5 className={styles.ClusterName}>{children}</h5>;
