+++ client/src/DeleteScriptButton.js
@@ -1,5 +1,7 @@
-import ConfirmedActionButton from './lib/ConfirmedActionButton';
-import { errorCode } from './lib/utils';
+import {
+  ConfirmedActionButton,
+  utils,
+} from './lib';
 import { useDeleteScript } from './api';
 import { useToast } from './ToastContext';
 
@@ -21,7 +23,7 @@ function DeleteScriptButton({
       } else {
         addToast(deleteFailedToast({
           script: script,
-          errorCode: errorCode(responseBody),
+          errorCode: utils.errorCode(responseBody),
         }));
       }
     } catch (e) {
